+++ src/main/events/misc/index.ts
@@ -9,6 +9,7 @@ import "./hydra-api-call";
 import "./install-common-redist";
 import "./install-hydra-decky-plugin";
 import "./is-main-window-open";
+import "./is-rclone-available";
 import "./open-checkout";
 import "./open-external";
 import "./open-main-window";
